src/main/java/zoidnet/dev/chat/service/UserService.java
package zoidnet.dev.chat.service;


import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.dao.DataAccessException;
import org.springframework.security.crypto.password.PasswordEncoder;
import org.springframework.stereotype.Service;
import zoidnet.dev.chat.controller.dto.UserDto;
import zoidnet.dev.chat.model.User;
import zoidnet.dev.chat.repository.UserRepository;


@Service
public class UserService {

    @Autowired
    private UserRepository userRepository;

    @Autowired
    private PasswordEncoder passwordEncoder;


    public void registerUser(UserDto userDto) throws DataAccessException {
        User newUser = userDto.toUser(passwordEncoder);
        userRepository.save(newUser);
    }

}
